pyproject.toml
[tool.poetry]
authors = ["Artem Malyshev <proofit404@gmail.com>"]
description = "Apply cruft tool continuously in pull requests."
name = "cruftbot"
version = "0.1.0"

[tool.poetry.dependencies]
Django = "3.1.7"
Jinja2 = "2.11.3"
dependencies = "6.0.1"
django-activity-stream = "0.10.0"
django-allauth = "0.44.0"
django-axes = "5.13.1"
django-cors-headers = "3.7.0"
django-csp = "3.7"
django-environ = "0.4.5"
django-feature-policy = "3.7.0"  # @todo #2 Setup django-feature-policy package.
django-filter = "2.4.0"  # @todo #2 Setup django-filter package.
django-guardian = "2.3.0"  # @todo #2 Setup django-guardian package.
django-health-check = "3.16.3"  # @todo #2 Setup django-health-check package.
django-http-referrer-policy = "1.1.1"  # @todo #2 Setup django-http-referrer-policy package.
django-model-utils = "4.1.1"
django-node-assets = "0.9.8"
django-reversion = "3.0.9"  # @todo #2 Setup django-reversion package.
django-silk = "4.1.0"  # @todo #2 Setup django-silk package.
django-split-settings = "1.0.1"
djangorestframework = "3.12.2"
drf-yasg = "1.20.0"
generics = "3.3.0"
psycopg2-binary = "2.8.6"
pydantic = "1.8.1"
pydantic-initialized = "0.4.0"
python = "3.9.2"
stories = "4.0.0"
stories-pydantic = "0.5.0"

[tool.poetry.dev-dependencies]
bandit = "1.7.0"
black = "20.8b1"
coverage = "5.4"
django-debug-toolbar = "3.2"
django-extensions = "3.1.1"
django-extra-checks = "0.8.0"
django-migration-linter = "2.5.1"
django-querycount = "0.7.0"
django-test-migrations = "1.1.0"
docformatter = "1.4"
flake8 = "3.9.0"
flake8-2020 = "1.6.0"
flake8-blind-except = "0.2.0"
flake8-bugbear = "20.11.1"
flake8-builtins = "1.5.3"
flake8-comprehensions = "3.3.1"
flake8-eradicate = "1.0.0"
flake8-executable = "2.1.1"
flake8-fixme = "1.1.1"
flake8-mutable = "1.2.0"
flake8-print = "4.0.0"
flake8-pytest = "1.3"
flake8-pytest-style = "1.3.0"
flake8-tidy-imports = "4.2.1"
ipython = "7.21.0"
mkdocs = "1.1.2"
pre-commit = "2.11.1"
pyupgrade = "2.10.1"
reorder-python-imports = "2.4.0"
stories-django = "0.1.0"
toml-sort = "0.18.0"
yamllint = "1.26.0"
